Migrated Identifiers components to TypeScript
