+++ src/main/webapp/ui/src/Inventory/components/Fields/Identifiers/AmberNumberField.tsx
@@ -33,4 +33,3 @@ const AmberNumberField: ComponentType<NumberFieldArgs> = withStyles<
 ));
 
 export default AmberNumberField;
-
