+++ src/main/webapp/ui/src/Inventory/components/Fields/Identifiers/PolygonCard.js.flow
@@ -1,0 +1,16 @@
+// @flow
+
+import React, { type Node, type ComponentType } from "react";
+import {
+  type GeoLocation,
+  type PolygonPoint,
+} from "../../../../stores/definitions/GeoLocation";
+import { type PolygonMessages } from "./GeoLocationField";
+
+type PolygonCardArgs = {|
+  editable: boolean,
+  geoLocation: GeoLocation,
+  doUpdateIdentifiers: () => void,
+|};
+
+declare export default ComponentType<PolygonCardArgs>;
