+++ src/main/webapp/ui/src/Inventory/components/Fields/Identifiers/PolygonDialog.tsx
@@ -1,5 +1,4 @@
-//@flow
-import React, { type Node, type ComponentType } from "react";
+import React, { type ReactNode, type ComponentType } from "react";
 import { observer } from "mobx-react-lite";
 import Button from "@mui/material/Button";
 import DialogActions from "@mui/material/DialogActions";
@@ -7,16 +6,16 @@ import DialogContent from "@mui/material/DialogContent";
 import Dialog from "@mui/material/Dialog";
 import DialogTitle from "@mui/material/DialogTitle";
 import FormControl from "@mui/material/FormControl";
-import PolygonCard from "./PolygonCard.js";
+import PolygonCard from "./PolygonCard";
 import { type GeoLocation } from "../../../../stores/definitions/GeoLocation";
 
-type PolygonDialogArgs = {|
-  open: boolean,
-  setOpen: (boolean) => void,
-  editable: boolean,
-  geoLocation: GeoLocation,
-  doUpdateIdentifiers: () => void,
-|};
+type PolygonDialogArgs = {
+  open: boolean;
+  setOpen: (open: boolean) => void;
+  editable: boolean;
+  geoLocation: GeoLocation;
+  doUpdateIdentifiers: () => void;
+};
 
 function PolygonDialog({
   open,
@@ -24,7 +23,7 @@ function PolygonDialog({
   editable,
   geoLocation,
   doUpdateIdentifiers,
-}: PolygonDialogArgs): Node {
+}: PolygonDialogArgs): ReactNode {
   const handleClose = () => {
     setOpen(false);
   };
@@ -50,4 +49,4 @@ function PolygonDialog({
   );
 }
 
-export default (observer(PolygonDialog): ComponentType<PolygonDialogArgs>);
+export default observer(PolygonDialog) as ComponentType<PolygonDialogArgs>;
