+++ src/main/webapp/ui/src/Inventory/components/Fields/Identifiers/__tests__/Identifiers.test.js.flow
@@ -1,0 +1,14 @@
+// @flow
+/* eslint-env jest */
+
+import React from "react";
+import { type InventoryRecord } from "../../../../../stores/definitions/InventoryRecord";
+import { IdentifiersList } from "../Identifiers";
+
+declare var describe: Function;
+declare var beforeEach: Function;
+declare var afterEach: Function;
+declare var test: Function;
+declare var expect: any;
+
+export const mockIGSNIdentifier = (recordType: "sample" | "container"): Object => ({});
