+++ src/main/webapp/ui/src/Inventory/components/Fields/Identifiers/__tests__/Identifiers.test.tsx
@@ -1,8 +1,6 @@
 /*
  * @jest-environment jsdom
  */
-//@flow
-/* eslint-env jest */
 import React from "react";
 import { render, cleanup } from "@testing-library/react";
 import "@testing-library/jest-dom";
